Add temporary logo and favicon

+++ setup.sh
@@ -4,7 +4,15 @@
 # TODO: Support systems with curl but not wget
 # TODO: Support systems with bsdtar but not tar
 
-wget https://github.com/jgm/pandoc/releases/download/3.6.2/pandoc-3.6.2-linux-amd64.tar.gz -O pandoc.tar.gz
-mkdir pandoc/
-tar -xf pandoc.tar.gz -C pandoc/ --strip-components=1
-rm pandoc.tar.gz
+mkdir -p bin/
+
+if [ ! -f bin/pandoc ]; then
+  wget https://github.com/jgm/pandoc/releases/download/3.6.2/pandoc-3.6.2-linux-amd64.tar.gz -O pandoc.tar.gz
+  tar -xf pandoc.tar.gz -C bin/ --strip-components=2 pandoc-3.6.2/bin/pandoc
+  rm pandoc.tar.gz
+fi
+
+if [ ! -f bin/magick ]; then
+  wget https://imagemagick.org/archive/binaries/magick -O bin/magick
+  chmod u+x bin/magick
+fi
